+++ database/migrations/2020_10_17_185147_create_services_table.php
@@ -22,7 +22,7 @@ class CreateServicesTable extends Migration
             $table->unsignedInteger('comission')->nullable();
             $table->boolean('conversion')->nullable();
 
-            $table->unsignedBigInteger('master_id')->nullable();
+            $table->unsignedInteger('master_id')->nullable();
 
             $table->timestamps();
             $table->softDeletes();
